requirements.txt
-i https://pypi.org/simple
aiosmtplib==2.0.2; python_version >= '3.7' and python_version < '4.0'
aiosqlite==0.17.0; python_version >= '3.6'
anyio==4.0.0; python_version >= '3.8'
async-timeout==4.0.3; python_full_version < '3.12.0'
asyncpg==0.29.0; python_full_version >= '3.8.0'
authlib==1.2.1
bcrypt==4.0.1; python_version >= '3.6'
blinker==1.7.0; python_version >= '3.8'
boto3==1.28.80; python_version >= '3.7'
botocore==1.31.80; python_version >= '3.7'
certifi==2023.7.22; python_version >= '3.6'
cffi==1.16.0; python_version >= '3.8'
chardet==5.2.0; python_version >= '3.7'
charset-normalizer==3.3.2; python_full_version >= '3.7.0'
click==8.1.7; python_version >= '3.7'
cryptography==41.0.5; python_version >= '3.7'
dnspython==2.4.2; python_version >= '3.8' and python_version < '4.0'
email-validator==1.3.1; python_version >= '3.5'
et-xmlfile==1.1.0; python_version >= '3.6'
exceptiongroup==1.1.3; python_version < '3.11'
fastapi==0.80.0; python_full_version >= '3.6.1'
fastapi-mail==1.2.5; python_version < '4.0' and python_full_version >= '3.8.1'
fillpdf==0.7.2; python_version >= '3.6'
h11==0.14.0; python_version >= '3.7'
httpcore==1.0.1; python_version >= '3.8'
httpx==0.25.1; python_version >= '3.8'
idna==3.4; python_version >= '3.5'
iso8601==1.1.0; python_version < '4.0' and python_full_version >= '3.6.2'
jinja2==3.1.2; python_version >= '3.7'
jmespath==1.0.1; python_version >= '3.7'
mangum==0.17.0; python_version >= '3.7'
markupsafe==2.1.3; python_version >= '3.7'
msal==1.25.0; python_version >= '2.7'
numpy==1.26.1; python_version < '3.11'
oauthlib==3.2.2; python_version >= '3.6'
openpyxl==3.1.2; python_version >= '3.6'
pandas==2.1.2; python_version >= '3.9'
passlib==1.7.4
pdf2image==1.16.3
pdfrw2==0.5.0
pillow==10.1.0; python_version >= '3.8'
pycparser==2.21
pydantic==1.9.1; python_full_version >= '3.6.1'
pyjwt==2.4.0; python_version >= '3.6'
pymupdf==1.23.6; python_version >= '3.8'
pymupdfb==1.23.6; python_version >= '3.8'
pypika-tortoise==0.1.6; python_version >= '3.7' and python_version < '4.0'
python-dateutil==2.8.2; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
python-dotenv==1.0.0; python_version >= '3.8'
python-http-client==3.3.7; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
python-multipart==0.0.6; python_version >= '3.7'
pytz==2023.3.post1
requests==2.31.0; python_version >= '3.7'
s3transfer==0.7.0; python_version >= '3.7'
sendgrid==6.10.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
sniffio==1.3.0; python_version >= '3.7'
starkbank-ecdsa==2.2.0
starlette==0.19.1; python_version >= '3.6'
tortoise-orm==0.19.2; python_version >= '3.7' and python_version < '4.0'
typing-extensions==4.8.0; python_version >= '3.8'
tzdata==2023.3; python_version >= '2'
urllib3==2.0.7; python_version >= '3.7'
uvicorn==0.24.0.post1; python_version >= '3.8'
xlsxwriter==3.1.9; python_version >= '3.6'
